Update project
- Make a fix for notification tab that's not full screen in main view tabbed.
- Fixed when clear a notification and undo it, it won't open correctly.
- Renamed from NotificationDialogBox to NotificationScaffold.
