+++ app/src/main/java/io/zoemeow/dutschedule/ui/view/main/NotificationScaffold.kt
@@ -1,18 +1,14 @@
-package io.zoemeow.dutschedule.ui.component.main.notification
+package io.zoemeow.dutschedule.ui.view.main
 
 import android.graphics.Bitmap
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.slideInVertically
 import androidx.compose.animation.slideOutVertically
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.rememberScrollState
@@ -26,11 +22,12 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarHost
 import androidx.compose.material3.SnackbarHostState
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateListOf
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -41,11 +38,13 @@ import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.unit.dp
 import io.zoemeow.dutschedule.R
 import io.zoemeow.dutschedule.model.NotificationHistory
+import io.zoemeow.dutschedule.ui.component.main.NotificationItem
 import io.zoemeow.dutschedule.utils.CustomDateUtil
+import okhttp3.internal.concurrent.TaskRunner
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun NotificationDialogBox(
+fun NotificationScaffold(
     isVisible: Boolean = false,
     itemList: List<NotificationHistory>,
     containerColor: Color,
@@ -56,7 +55,6 @@ fun NotificationDialogBox(
     onClick: ((NotificationHistory) -> Unit)? = null,
     onClear: ((NotificationHistory) -> Unit)? = null,
     onClearAll: (() -> Unit)? = null,
-    height: Float = 0.7f,
     opacity: Float = 1f,
 ) {
     AnimatedVisibility(
@@ -112,8 +110,7 @@ fun NotificationDialogBox(
                 content = {
                     Column(
                         modifier = Modifier
-                            .fillMaxWidth()
-                            .fillMaxHeight(height)
+                            .fillMaxSize()
                             .padding(it)
                             .padding(horizontal = 15.dp)
                     ) {
@@ -141,6 +138,7 @@ fun NotificationDialogBox(
                                         group.value.forEach { item ->
                                             NotificationItem(
                                                 modifier = Modifier.padding(top = 2.dp, bottom = 5.dp),
+                                                isVisible = true,
                                                 opacity = opacity,
                                                 onClick = { onClick?.let { it(item) } },
                                                 onClear = { onClear?.let { it(item) } },
